pairs_trading/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
soupsieve==2.2.1
statsmodels==0.12.2
urllib3==1.26.4
yfinance==0.1.59
